Don't fail on missing credentials; add gradle.properties
OSSRH credentials are determined using the inbuilt `PasswordCredentials`
class and the curse API key does not error when missing.

The gradle.properties file is now included in the repo so that the
kotlin code style and other official things can be set there. Please
store credentials in environment variables, `~/.gradle` or pass them as
command line arguments.(See
https://docs.gradle.org/current/samples/sample_publishing_credentials.html)